+++ packages/itables_anywidget/js/widget.ts
@@ -19,7 +19,7 @@ function render({ model, el }: RenderContext<WidgetModel>) {
 	el.appendChild(table);
 
 	function update_classes() {
-		table.className = model.get("classes").split(" ");
+		table.className = model.get("classes");
 	}
 	function update_style() {
 		table.setAttribute('style', model.get("style"));