+++ tests/test_extension_arguments.py
@@ -1,8 +1,6 @@
-import pandas as pd
 import pytest
 
-from itables import JavascriptCode
-from itables.javascript import get_itables_extension_arguments, pd_style
+from itables.javascript import get_itables_extension_arguments
 
 
 def test_get_itables_extension_arguments(df):
@@ -12,8 +10,9 @@ def test_get_itables_extension_arguments(df):
         pytest.skip(str(e))
 
     assert set(dt_args) <= {
+        "table_html",
         "data_json",
-        "columns",
+        "column_filters",
         "layout",
         "order",
         "text_in_header_can_be_selected",
@@ -21,7 +20,7 @@ def test_get_itables_extension_arguments(df):
         "downsampling_warning",
     }, set(dt_args)
     assert isinstance(dt_args["data_json"], str)
-    assert isinstance(dt_args["columns"], list)
+    assert isinstance(dt_args["table_html"], str)
 
     assert set(other_args) <= {
         "classes",
@@ -32,38 +31,3 @@ def test_get_itables_extension_arguments(df):
     assert isinstance(other_args["classes"], str)
     assert isinstance(other_args["style"], str)
     assert other_args["caption"] is None
-
-
-def test_no_use_to_html():
-    with pytest.raises(
-        TypeError,
-        match="In Python apps, these options are not available",
-    ):
-        get_itables_extension_arguments(pd.DataFrame({"a": [0]}), use_to_html=True)
-
-
-def test_no_javascript_code():
-    with pytest.raises(
-        TypeError, match="Javascript code can't be passed to the extension"
-    ):
-        get_itables_extension_arguments(
-            pd.DataFrame({"a": [0]}),
-            columnDefs=[
-                {
-                    "targets": "_all",
-                    "render": JavascriptCode(
-                        "$.fn.dataTable.render.number(',', '.', 3, '$')"
-                    ),
-                }
-            ],
-        )
-
-
-def test_no_style_object():
-    if pd_style is None:
-        pytest.skip("Pandas Style is not available")
-    with pytest.raises(
-        NotImplementedError,
-        match="Pandas style objects can't be used with the extension",
-    ):
-        get_itables_extension_arguments(pd.DataFrame({"a": [0]}).style)